requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
Pillow==10.0.1
python-pptx==0.6.21
reportlab==4.0.4
gunicorn==21.2.0
